move TypeBound to types.rs
