+++ src/types.rs
@@ -90,6 +90,57 @@ impl TypeTag {
     }
 }
 
+use itertools::FoldWhile::{Continue, Done};
+use itertools::Itertools;
+use serde::{Deserialize, Serialize};
+
+#[derive(Copy, Clone, PartialEq, Eq, Hash, Debug, derive_more::Display, Serialize, Deserialize)]
+/// Bounds on capabilities of a type.
+pub enum TypeBound {
+    /// The equality operation is valid on this type.
+    #[serde(rename = "E")]
+    Eq,
+    /// The type can be copied in the program.
+    #[serde(rename = "C")]
+    Copyable,
+}
+
+impl TypeBound {
+    /// Returns the smallest TypeTag containing both the receiver and argument.
+    /// (This will be one of the receiver or the argument.)
+    pub fn union(self, other: Self) -> Self {
+        if self.contains(other) {
+            self
+        } else {
+            debug_assert!(other.contains(self));
+            other
+        }
+    }
+
+    /// Report if this bound contains another.
+    pub fn contains(&self, other: TypeBound) -> bool {
+        use TypeBound::*;
+        match (self, other) {
+            (Copyable, Eq) => true,
+            (Eq, Copyable) => false,
+            _ => true,
+        }
+    }
+}
+
+/// Calculate the least upper bound for an iterator of bounds
+fn least_upper_bound(mut tags: impl Iterator<Item = Option<TypeBound>>) -> Option<TypeBound> {
+    tags.fold_while(Some(TypeBound::Eq), |acc, new| {
+        if let (Some(acc), Some(new)) = (acc, new) {
+            Continue(Some(acc.union(new)))
+        } else {
+            // if any type is unbounded, short-circuit
+            Done(None)
+        }
+    })
+    .into_inner()
+}
+
 #[cfg(test)]
 mod test {
     use crate::{ops::AliasDecl, types::type_enum::Type};
